Leave import receipts specific to op mainnet
