+++ bin/reth/src/cli/mod.rs
@@ -152,9 +152,6 @@ impl<Ext: clap::Args + fmt::Debug> Cli<Ext> {
             Commands::Init(command) => runner.run_blocking_until_ctrl_c(command.execute()),
             Commands::InitState(command) => runner.run_blocking_until_ctrl_c(command.execute()),
             Commands::Import(command) => runner.run_blocking_until_ctrl_c(command.execute()),
-            Commands::ImportReceipts(command) => {
-                runner.run_blocking_until_ctrl_c(command.execute())
-            }
             #[cfg(feature = "optimism")]
             Commands::ImportOp(command) => runner.run_blocking_until_ctrl_c(command.execute()),
             #[cfg(feature = "optimism")]
@@ -197,9 +194,6 @@ pub enum Commands<Ext: clap::Args + fmt::Debug = NoArgs> {
     /// This syncs RLP encoded blocks from a file.
     #[command(name = "import")]
     Import(import::ImportCommand),
-    /// This imports RLP encoded receipts from a file.
-    #[command(name = "import-receipts")]
-    ImportReceipts(import_receipts::ImportReceiptsCommand),
     /// This syncs RLP encoded OP blocks below Bedrock from a file, without executing.
     #[cfg(feature = "optimism")]
     #[command(name = "import-op")]
