+++ bin/reth/src/commands/mod.rs
@@ -6,7 +6,6 @@ pub mod debug_cmd;
 pub mod dump_genesis;
 pub mod import;
 pub mod import_op;
-pub mod import_receipts;
 pub mod import_receipts_op;
 
 pub mod init_cmd;
